feature(gif): employees can comment on a gif post
[Fixes #12]
